WebApplication_sourcecode/requirements.txt
gunicorn==20.1.0
colorama==0.4.4
flask==2.0.2
Pillow
numpy
tensorflow-cpu==2.8.0

